Uses ExifInterface from Support Library and saves output rotated without EXIF
The Support Library was updated to 25.1.0, which includes a support ExifInterface.

Using ExifInterface from Support Library allows us to read the EXIF metadata from
any Uri (protocol independent, using ContentResolver.openInputStream).

For the same reason (to be protocol independent), the EXIF information is not copied
to the output anymore. Instead, the cropped image is rotated using Matrix.postRotate
(as a result of this, it is not necessary to rewrite the output just to copy the EXIF
rotation as done before).